Remove inclusion of the ClangSACheckers.h header
The `ClangSACheckers.h` header is clang-internal, e.g. it is used for
clang-internal stuff but is not installed in the include directory.
Hence, it is not availible if llvm is not build from source.
However, it is apparently not required for compilation of the (external)
checker module. Removing it should be safe.